refactor: move message definitions to stubs
