+++ venom/rpc/reflect/openapi.py
@@ -1,8 +1,11 @@
 from itertools import groupby
 from collections import defaultdict
 from venom.rpc.reflect.reflect import Reflect
+from venom.rpc.reflect.stubs import TypeMessage, MethodMessage, \
+    ParameterMessage, ResponsesMessage, ResponseMessage, \
+    FieldsMessage, InfoMessage, OpenAPISchema
 from venom.protocol import JSON
-from venom.fields import Field, Repeat, RepeatField, String, Map, Bool
+from venom.fields import Field, RepeatField
 from venom import Message
 from venom.message import field_names, fields
 from venom.rpc.method import Method, HTTPFieldLocation
@@ -33,57 +36,6 @@ BODY_PARAMETER = {
 QUERY_PARAMETER = {
     'is_in': 'query',
 }
-
-
-class TypeMessage(Message):
-    type = String()  # TODO: enum
-    items = Field('venom.rpc.reflect.openapi.TypeMessage')
-    ref = String(name='$ref')
-
-
-class FieldsMessage(Message):
-    type = String()
-    properties = Map(Field(TypeMessage))
-    ref = String(name='$ref')
-
-
-class ParameterMessage(Message):
-    is_in = String(name='in')  # TODO: enum
-    required = Bool()
-    name = String()
-    type = String()
-    items = Field(TypeMessage)
-    schema = Field(FieldsMessage)
-
-
-class ResponseMessage(Message):
-    description = String()
-    schema = Field(FieldsMessage)
-
-
-class ResponsesMessage(Message):
-    default = Field(ResponseMessage)  # TODO: error codes
-
-
-class MethodMessage(Message):
-    produces = Repeat(String())  # TODO: default value for RepeatField
-    responses = Field(ResponsesMessage)
-    parameters = Repeat(ParameterMessage)
-
-
-class InfoMessage(Message):
-    version = String(default='0.0.1')  # TODO: version
-    title = String()
-
-
-class OpenAPISchema(Message):
-    swagger = String(default='2.0')
-    schemes = Repeat(String())  # TODO: default value for RepeatField
-    consumes = Repeat(String())  # TODO: default value for RepeatField
-    produces = Repeat(String())  # TODO: default value for RepeatField
-    info = Field(InfoMessage)
-    paths = Map(Map(Field(MethodMessage)))
-    definitions = Map(Field(FieldsMessage))
 
 
 def field_type(field: Field) -> str:
